feat(pdv): Implement initial PDV schema and related tables
- Create tables for PDV functionality including pdv_terminal, pdv_caixa, pdv_caixa_mov, pdv_fiado, pdv_fiado_pagamento, invoice_payment, pdv_venda_suspensa, and pdv_audit_log.
- Modify existing tables: product_tbl, invoice_tbl, invoice_details, customer_tbl, user, and product_return to support new PDV features.
- Add initial data for PDV menu and roles.

feat(pdv): Enhance fiado functionality in PDV

- Modify pdv_caixa_mov to include new transaction types for fiado.
- Add caixa_id to pdv_fiado_pagamento for linking payments to cash registers.
- Introduce a computed saldo column in pdv_fiado for easier queries.

feat(pdv): Add return details for product exchanges

- Extend return_details table with invoice_detail_id and motivo columns for better tracking of returns.
- Create an index for faster lookups on returned quantities.

feat(pdv): Finalize product and user form fields

- Add unique index on codigo_balanca in product_tbl.
- Ensure ean_gtin is unique by replacing the existing non-unique index with a unique one.

+++ inventro_mainfile/docker/mysql/init/01-schema.sql
@@ -195,12 +195,12 @@ CREATE TABLE IF NOT EXISTS `invoice_details` (
   `invoice_id` varchar(20) COLLATE utf8_unicode_ci NOT NULL,
   `invoice_details_id` varchar(20) COLLATE utf8_unicode_ci NOT NULL,
   `product_id` varchar(20) COLLATE utf8_unicode_ci NOT NULL,
-  `quantity` int(11) NOT NULL,
-  `price` decimal(10,0) DEFAULT NULL,
-  `discount` decimal(10,0) DEFAULT NULL,
-  `total_price` decimal(10,0) DEFAULT NULL,
-  `discount_amount` decimal(10,0) DEFAULT NULL,
-  `tax` decimal(10,0) DEFAULT NULL,
+  `quantity` decimal(10,3) NOT NULL,
+  `price` decimal(10,2) DEFAULT NULL,
+  `discount` decimal(10,2) DEFAULT NULL,
+  `total_price` decimal(10,2) DEFAULT NULL,
+  `discount_amount` decimal(10,2) DEFAULT NULL,
+  `tax` decimal(10,2) DEFAULT NULL,
   `created_date` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
